+++ judo-runtime-core-jsl-itest/pom.xml
@@ -88,6 +88,20 @@
 
                 </executions>
             </plugin>
+
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <version>3.10.1</version>
+                <configuration>
+                    <source>11</source>
+                    <target>11</target>
+                    <compilerArgs>
+                        <arg>-AreportPath=${basedir}/target/classes/requirements-report.csv</arg>
+                    </compilerArgs>
+                </configuration>
+            </plugin>
+
             <!--
             <plugin>
                 <groupId>org.apache.felix</groupId>
@@ -103,31 +117,6 @@
                 </configuration>
             </plugin>
             -->
-
-            <!--plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.8.0</version>
-                <configuration>
-                    <annotationProcessors>
-                        <annotationProcessor>hu.blackbelt.judo.annotation.RequirementProcessor</annotationProcessor>
-                    </annotationProcessors>
-                    <debug>true</debug>
-                    <optimize>true</optimize>
-                    <compilerArguments>
-                        <AaddGeneratedAnnotation>true</AaddGeneratedAnnotation>
-                        <Adebug>true</Adebug>
-                    </compilerArguments>
-                    <annotationProcessorPaths>
-                        <annotationProcessorPath>
-                            <groupId>hu.blackbelt.judo.annotation</groupId>
-                            <artifactId>judo-annotations</artifactId>
-                            <version>1.0-SNAPSHOT</version>
-                        </annotationProcessorPath>
-                    </annotationProcessorPaths>
-                </configuration>
-
-            </plugin-->
         </plugins>
     </build>
 
@@ -178,8 +167,15 @@
         </dependency>
 
         <dependency>
-            <groupId>hu.blackbelt.judo.annotation</groupId>
-            <artifactId>judo-annotations</artifactId>
+            <groupId>hu.blackbelt.judo.requirement.report</groupId>
+            <artifactId>judo-requirement-report-annotation</artifactId>
+            <version>1.0-SNAPSHOT</version>
+            <scope>compile</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>hu.blackbelt.judo.requirement.report</groupId>
+            <artifactId>judo-requirement-report-processor</artifactId>
             <version>1.0-SNAPSHOT</version>
             <optional>true</optional>
             <scope>compile</scope>
